Fix so assuming role works when the current IAM principal is an assumed role
The minimal test case is:

    $ ./assume-role --role arn:aws:iam::000000000001:role/abc ./assume-role --role arn:aws:iam::000000000000:role/def
    ERROR: unable to get username from AWS: ValidationError: Must specify userName when calling with non-User credentials
            status code: 400, request id: 00000000-0000-0000-0000-000000000000

The real-world use case is using assume-role from an instance, where the IAM
principal is the role assigned to the instance.

Since assumed roles don't have an username, the session name for the assumed role
needs to be passed as a command-line parameter, which is mandatory when the current
principal is an assumed role.